Moved helper functions to service folder
