+++ src/screens/Forum/CreateForumScreen.js
@@ -6,7 +6,7 @@ import styled from "styled-components/native";
 
 import { firebase } from "../../services/Firebase/Config";
 import { useNavigation } from "@react-navigation/native";
-import { createForum } from "../../components/Forum/HandleForum";
+import { createForum } from "../../services/Forum/HandleForum";
 const imageType = {
   PROFILE: 0,
   BANNER: 1
